+++ src/daf/messagedata/__init__.py
@@ -2,3 +2,5 @@ from .basedata import *
 from .textdata import *
 from .voicedata import *
 from .dynamicdata import *
+
+from .file import *